api/requirements.txt
asgiref==3.8.1
attrs==23.2.0
black==24.1.1
CacheControl==0.14.0
cachetools==5.3.3
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cryptography==42.0.5
Django==5.0.6
django-autoslug==1.9.9
django-cors-headers==4.3.1
django-extensions==3.2.3
django-filter==23.5
django-js-asset==2.2.0
django-mptt==0.16.0
django-phonenumber-field==7.3.0
django-sql-sniffer==1.0.3
djangorestframework==3.14.0
drf-spectacular==0.27.1
environs==10.3.0
firebase-admin==6.5.0
google-api-core==2.18.0
google-api-python-client==2.125.0
google-auth==2.29.0
google-auth-httplib2==0.2.0
google-cloud-core==2.4.1
google-cloud-firestore==2.16.0
google-cloud-storage==2.16.0
google-crc32c==1.5.0
google-resumable-media==2.7.0
googleapis-common-protos==1.63.0
grpcio==1.62.1
grpcio-status==1.62.1
gunicorn==21.2.0
httplib2==0.22.0
idna==3.6
inflection==0.5.1
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
Markdown==3.5.1
marshmallow==3.20.2
msgpack==1.0.8
mypy-extensions==1.0.0
packaging==23.2
pathspec==0.12.1
phonenumbers==8.13.35
platformdirs==4.2.0
proto-plus==1.23.0
protobuf==4.25.3
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
Pygments==2.17.2
PyJWT==2.8.0
pyparsing==3.1.2
python-dotenv==1.0.1
pytz==2023.3.post1
PyYAML==6.0.1
referencing==0.32.1
requests==2.31.0
rpds-py==0.17.1
rsa==4.9
sqlparse==0.5.0
tomli==2.0.1
typing_extensions==4.11.0
uritemplate==4.1.1
urllib3==2.2.1
whitenoise==6.6.0
